gjfs

+++ src/tctlab/functs.py
@@ -52,7 +52,7 @@ class waveforms:
         ADC = 3.1422522482546027
         
         for i in range(len(wfms[0])):
-            slope_seg = ped[slopes][sw[i]*64 : sw[i]*64 + npts]
+            slope_seg = ped[slopes][int(sw[i]*64) : int(sw[i]*64 + npts)]
             
             data[:,i] = wfms[:,i] - ( slope_seg * wfms[:,i] / ADC)
